ref:(prism): ensure child pkgs not dependent on parents (#554)
